+++ FastAnimations/Framework/ModConfig.cs
@@ -8,5 +8,8 @@ namespace Pathoschild.Stardew.FastAnimations.Framework
         *********/
         /// <summary>Whether to check for updates to the mod.</summary>
         public bool CheckForUpdates { get; set; } = true;
+
+        /// <summary>Whether to instantly eat or drink food.</summary>
+        public bool InstantEat { get; set; } = true;
     }
 }