+++ FastAnimations/release-notes.md
@@ -3,3 +3,4 @@
 # Release notes
 ## 1.0
 * Initial version.
+* Added support for instant eating & drinking.